backend/requirements.txt
flask
flask_cors
transformers
transformers[torch]
datasets
scipy
openai-whisper
temp